General Update
You can now get the number of elements left in the queue and get the
number of elements on screen.
You can also specify if an element must be shown in which case if stop
is called whilst the element before is on this will still appear and you
can specify if an element should not be shown if it would be the first
one to scroll. Both these features are useful if you want to add
separators.